Implement franchise fork / multi-source support
Replaces the single save_file_path/league_guid on franchises with a
franchise_sources table, allowing a franchise to span multiple save game
files when the user exports to a new league in SMB4.

Schema changes:
- registry: add franchise_sources (id, franchise_id, save_file_path,
  league_guid, season_offset, added_at); remove save_file_path/league_guid
  from franchises
- companion: seasons now uses AUTOINCREMENT id with (league_guid,
  save_game_season_id) as uniqueness key; adds player_alt_guids and
  team_alt_guids tables for cross-fork GUID re-association

Import pipeline:
- season_num stored at import time = save game season num + source offset
- UpsertPlayer uses 3-tier lookup: primary GUID → alt GUIDs → fuzzy match
  (name + bat/throw hand + chemistry), registering new alt GUIDs on match
- UpsertTeam uses same 3-tier approach with name-match fallback

App bindings:
- SetInitialSource / AddFranchiseSource / ReplaceActiveFranchiseSource
  replace SetFranchiseSaveFile
- SyncSeason reads from the highest-offset source automatically
- FranchiseDTO now carries hasActiveSource and activeSourcePath

+++ internal/store/player_query.go
@@ -168,7 +168,7 @@ func (s *PlayerQueryStore) GetPlayerSeasonLog(ctx context.Context, playerID int6
 // season_id to slice index (used by the caller to merge playoff stats).
 func (s *PlayerQueryStore) scanSeasonLogRows(
 	ctx context.Context, playerID int64, isRegularSeason int,
-) ([]models.PlayerSeasonLogRow, map[int]int, error) {
+) ([]models.PlayerSeasonLogRow, map[int64]int, error) {
 	rows, err := s.db.QueryContext(ctx, `
 SELECT
     s.id                              AS season_id,
@@ -222,7 +222,7 @@ ORDER BY s.season_num ASC
 	defer func() { _ = rows.Close() }()
 
 	var out []models.PlayerSeasonLogRow
-	index := map[int]int{}
+	index := map[int64]int{}
 
 	for rows.Next() {
 		var row models.PlayerSeasonLogRow
